pyproject.toml
[tool.poetry]
name = "bot"
version = "0.1.0"
description = "A personal discord bot for servers I like."
authors = ["du-cki <qwuakers@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
asyncpg = "^0.27.0"
toml = "^0.10.2"
cachetools = "^5.2.0"
yt-dlp = "^2023.7.6"
beautifulsoup4 = "^4.11.2"
lxml = "^4.9.2"
pillow = "^9.4.0"
psutil = "^5.9.4"
pynacl = "^1.5.0"
pygit2 = "^1.12.1"
jishaku = {git = "https://github.com/Gorialis/jishaku"} 
discord-py = {git = "https://github.com/Rapptz/discord.py"}
cloudscraper = "^1.2.71"

[tool.poetry.group.dev.dependencies]
black = "^22.10.0"
asyncpg-stubs = "^0.27.0"

[tool.pyright]
pythonVersion = "3.11"
typeCheckingMode = "strict"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
